added features

+++ AzureDevOps/Add_Remove_Agents.ps1
@@ -1,19 +1,13 @@
 param (
   [string]$rootPath = "c:\",
-  [string]$RemoveAzureURL = "https://adam-dc01/DefaultCollection/",
-  [string]$InstallAzureURL = "https://adam-dc01/DefaultCollection/",
+  [string]$RemoveAzureURL,
+  [string]$InstallAzureURL,
   [string]$Pass,
   [string]$user,
   [string]$AzurePool,
-  [string]$pat
+  [string]$pat,
+  [string]$Check="true"
 )
-#TOkEN:g47wmt6neeeydu5rc7rnqeuvj5ozylojcwhy7tsh4mzjki4ekr4q
-<#
-#need to do:
-- remoge with negotiate and isntall PAT
-- install new Agent
-# check about azure devops agent
-#>
 function SelectionMenu {
   param
   (
@@ -38,6 +32,7 @@ function SelectionMenu {
   Write-Host "4: Remove And Re-install -- Selected"
   Write-Host "5: Remove And Re-install New Agents (clean install)"
   Write-Host "6: install New Agents (clean install)"
+  Write-Host "7: Only Download"
   Write-Host "Q: Press 'Q' to quit." -foreground red
 }
 function MultiSelection{
@@ -52,8 +47,6 @@ function MultiSelection{
       # Default choice indexes (i.e. selected on [Enter])
       [int[]]$DefaultChoice = @(0)
     )
-    if ($args) { throw "Unknown parameters: $args" }
-
     $descriptions = @()
     for($i = 0; $i -lt $Choices.Count; $i++) {
       $c = [System.Management.Automation.Host.ChoiceDescription]("$($Choices[$i]) &$($i+1)")
@@ -70,20 +63,34 @@ function NewAgent {
     [string]$RootPath="c:\",
     [string]$NewAgentDirectory,
     [int]$NewAgentsAmount,
-    [string]$Download= "https://vstsagentpackage.azureedge.net/agent/2.168.2/vsts-agent-win-x64-2.168.2.zip"
+    [string]$Download= "https://vstsagentpackage.azureedge.net/agent/2.168.2/vsts-agent-win-x64-2.168.2.zip",
+    [string]$install="true"
   )
-  Invoke-WebRequest -Uri $Download -OutFile "$HOME\Downloads\vsts-agent.zip"
-  for($i = 0; $i -lt $NewAgentsAmount; $i++) {
-    $number=$i+1
-    $NewAgent="${NewAgentDirectory}0" + $number
-    if (Test-Path "$RootPath$NewAgent") {
-      Remove-Item -Path "$RootPath$NewAgent" -Recurse -Force
+  $AgentVersion=$Download.Split("/") | Where-Object{$_ -like "*.zip"}
+  if (Test-Path -Path "$HOME\Downloads\${AgentVersion}") {
+    Write-Host "Agnet allready Downloaded"
+  }
+  else {
+    Write-Host "Downloading Agent from $Download" -ForegroundColor Yellow
+    [Net.ServicePointManager]::SecurityProtocol = [Net.SecurityProtocolType]::Tls12
+    Invoke-WebRequest -Uri $Download -OutFile "$HOME\Downloads\${AgentVersion}" 
+  }
+  if ($install -like "true") {
+    for($i = 0; $i -lt $NewAgentsAmount; $i++) {
+      $number=$i+1
+      $NewAgent="${NewAgentDirectory}0" + $number
+      Write-Host "Creating Agnet $NewAgent" -ForegroundColor Yellow
+      if (Test-Path "$RootPath$NewAgent") {
+        Remove-Item -Path "$RootPath$NewAgent" -Recurse -Force
+        Write-Host "Removing Directory of $NewAgent" -ForegroundColor Yellow
+      }
+      Write-Host "Creating Directory for $NewAgent" -ForegroundColor Yellow
+      Set-Location $RootPath
+      mkdir $NewAgent ; Set-Location $NewAgent
+      Add-Type -AssemblyName System.IO.Compression.FileSystem ; [System.IO.Compression.ZipFile]::ExtractToDirectory("$HOME\Downloads\${AgentVersion}", "$PWD")
     }
     Set-Location $RootPath
-    mkdir $NewAgent ; Set-Location $NewAgent
-    Add-Type -AssemblyName System.IO.Compression.FileSystem ; [System.IO.Compression.ZipFile]::ExtractToDirectory("$HOME\Downloads\vsts-agent.zip", "$PWD")
   }
-  Set-Location $RootPath
 }
 function InstallRemoveAgent {
   param (
@@ -100,10 +107,10 @@ function InstallRemoveAgent {
     [string]$NewAgentPath
   )
   if ($method -like "remove-negotiate") {
-    & $agentsList[$i].fullname remove --unattended --url $RemoveAzureURL --auth negotiate --userName $user --password $Pass --pool $AzurePool --agent $AgentName --runAsService --runAsAutoLogon  
+    & $agentsList[$i].fullname remove --unattended --url $RemoveAzureURL --auth negotiate --userName $user --password $Pass
   }
   elseif ($method -like "remove-pat") {
-    & $agentsList[$i].fullname remove --unattended --url $RemoveAzureURL --auth pat --token $token --pool $AzurePool --agent $AgentName --runAsService --runAsAutoLogon  
+    & $agentsList[$i].fullname remove --unattended --url $RemoveAzureURL --auth pat --token $token
   }
   elseif ($method -like "install-negotiate") {
     & $agentsList[$i].fullname --unattended --url $InstallAzureURL --auth negotiate --userName $user --password $Pass --pool $AzurePool --agent $AgentName --runAsService --runAsAutoLogon  
@@ -112,22 +119,24 @@ function InstallRemoveAgent {
     & $agentsList[$i].fullname  --unattended --url $InstallAzureURL --auth pat --token $token --pool $AzurePool --agent $AgentName --runAsService --runAsAutoLogon  
   }
   elseif ($method -like "install-pat-clean") {
-    & $NewAgentPath --unattended --url $InstallAzureURL --auth pat --token $token --pool $AzurePool --agent $AgentName --runAsService --runAsAutoLogon  
+    & $NewAgentPath --sslskipcertvalidation --unattended --url $InstallAzureURL --auth pat --token $token --pool $AzurePool --agent $AgentName --runAsService --runAsAutoLogon  
   }
 }
 #get all agents installed on C:
-$ScanComputer=Get-ChildItem -recurse c:\ -Directory -Depth 0 | Where-Object {($_.name -notmatch "Program Files|Users|Windows")}
-$agentsList=@()
-foreach ($scan in $ScanComputer) {
-  Write-Host "Scanning"$scan.Name"..." -ForegroundColor yellow
-  $findConfigs=Get-ChildItem -Path $scan.FullName -Include *.cmd -file -recurse -Filter config.cmd
-  if ($null -ne $findConfigs) {
-    foreach ($findConfig in $findConfigs) {
-      $outputs=get-ChildItem -Path $findConfig.PSParentPath -Filter *.cmd
-      foreach ($output in $outputs) {
-        $compare=$agentsList | where-object {$_.DirectoryPath -like $output.Directory.fullname}
-        if (($outputs.Count -eq 2) -and ($null -eq $compare)) {
-          $agentsList+=[pscustomobject]@{Agent=$output.Directory.Name;FullName=$output.FullName;DirectoryPath=$output.Directory.fullname}
+if ($Check -like "True") {
+  $ScanComputer=Get-ChildItem -recurse $rootPath -Directory -Depth 0 | Where-Object {($_.name -notmatch "Program Files|Users|Windows")}
+  $agentsList=@()
+  foreach ($scan in $ScanComputer) {
+    Write-Host "Scanning"$scan.Name"..." -ForegroundColor yellow
+    $findConfigs=Get-ChildItem -Path $scan.FullName -Include *.cmd -file -Filter config.cmd -Depth 0
+    if ($null -ne $findConfigs) {
+      foreach ($findConfig in $findConfigs) {
+        $outputs=get-ChildItem -Path $findConfig.PSParentPath -Filter *.cmd
+        foreach ($output in $outputs) {
+          $compare=$agentsList | where-object {$_.DirectoryPath -like $output.Directory.fullname}
+          if (($outputs.Count -eq 2) -and ($null -eq $compare)) {
+            $agentsList+=[pscustomobject]@{Agent=$output.Directory.Name;FullName=$output.FullName;DirectoryPath=$output.Directory.fullname}
+          }
         }
       }
     }
@@ -155,11 +164,14 @@ switch ($input1) {
   }'6'{
     Write-Host "install New Agents (clean install)"
     $ActionSelection = "install-clean"
+  }'7'{
+    Write-Host "Only Download Agent"
+    $ActionSelection = "download"
   }'q' {
       return
   }
 }
-while (($input1 -gt 6) -or ($input1 -lt 1)) {
+while (($input1 -gt 7) -or ($input1 -lt 1)) {
   $input1 = Read-Host "Please select The Output Format"
   switch ($input1) {
     '1' {
@@ -180,6 +192,9 @@ while (($input1 -gt 6) -or ($input1 -lt 1)) {
     }'6'{
       Write-Host "install New Agents (clean install)"
       $ActionSelection = "install-clean"
+    }'7'{
+      Write-Host "Only Download Agent"
+      $ActionSelection = "download"
     }'q' {
         return
     }
@@ -292,4 +307,7 @@ elseif ($ActionSelection -like "install-clean") {
     $NewAgentPath= "$rootPath$NewAgentDirectory"+"0"+$number+"\config.cmd"
     InstallRemoveAgent -method "install-pat-clean" -NewAgentPath $NewAgentPath -InstallAzureURL $InstallAzureURL -token $pat -AzurePool $AzurePool -AgentName $AgentName
   }
+}
+elseif ($ActionSelection -like "download") {
+  NewAgent -install "false"
 }
